+++ example/lib/pages/others/example_bottom_main.dart
@@ -10,7 +10,8 @@ class ExampleMain extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Padding(
-        padding: const EdgeInsets.all(8) /* padding of all direction + bottomSafeArea */
+        padding: const EdgeInsets.all(
+                8) /* padding of all direction + bottomSafeArea */
             .copyWith(bottom: 8 + MediaQuery.paddingOf(context).bottom),
         child: HalfActionButtonGrid(
           buttons: pages.map((page) => page.getItemWidget(context)).toList(),