+++ lib/flutter_naver_map.dart
@@ -16,6 +16,7 @@ import 'package:flutter_naver_map/src/messaging/messaging.dart';
 import 'package:flutter_naver_map/src/util/image_util.dart';
 import 'package:flutter_naver_map/src/util/math.dart';
 import 'package:flutter_naver_map/src/util/widget_to_image.dart';
+import "package:flutter_naver_map/src/widget/elements/naver_logo_widget.dart";
 import "package:meta/meta.dart";
 
 /*
@@ -46,9 +47,11 @@ part 'src/exceptions/exceptions.dart';
   --- initializer ---
 */
 
-part 'src/initializer/flutter_naver_map_sdk_initializer.dart';
+part 'src/initializer/flutter_naver_map_initializer.dart';
 
-part 'src/initializer/flutter_naver_map_sdk_initializer_impl.dart';
+part 'src/initializer/flutter_naver_map_legacy_initializer.dart';
+
+part 'src/initializer/flutter_naver_map_legacy_initializer_impl.dart';
 /*
   --- type ---
 */
@@ -141,4 +144,4 @@ part 'src/widget/map_widget.dart';
 
 part 'src/widget/platform_view.dart';
 
-part 'src/widget/control_widget/zoom_control_widget.dart';
+part 'src/widget/elements/zoom_control_widget.dart';